move folder and url functionality from computation.py to folder.py
